+++ meta/manifest.lua
@@ -1,0 +1,7 @@
+-- meta/manifest.lua
+-- Repository systems listing
+return {
+  systems = {
+    "apps/cc-deploy/meta/manifest.lua"
+  }
+}